+++ packages/mtbird-shared/src/types/Component.ts
@@ -29,6 +29,14 @@ export interface IVariable {
 
 export type EventAction = 'click' | 'hover' | 'blur' | 'dbclick' | 'scroll';
 export type EventType = 'link' | 'link-blank' | 'submit' | 'clear' | 'open-modal' | 'close-modal' | 'inline-code' | 'change-variable';
+export type SearchOperator =
+  | 'eq' // 等于
+  | 'lt' // 小于
+  | 'gt' // 大于
+  | 'le' // 小于等于
+  | 'ge' // 大于等于
+  | 'ne' // 不等于
+  | 'like'; // 模糊匹配
 
 export interface IEvent {
   action?: EventAction;
@@ -56,8 +64,8 @@ export interface ISearch {
   // keyPath or field id
   keyPath: string;
   display: string;
-  operator: 'equals' | 'not-equals';
-  default: boolean;
+  operator: SearchOperator;
+  defaultValue: boolean;
   value: string;
 }
 
@@ -84,6 +92,7 @@ export interface IComponentInstance {
   props: IProps;
   layout?: 'absolute' | 'grid' | 'flex';
   data?: {
+    isDataContainer?: boolean; // 是否是数据容器（具备数据查询和暂存功能的组件）
     title?: string;
     alias?: string; // 组件自定义显示别名
     showIcon?: string; // 展示相关，某些组件是否展示icon
@@ -106,7 +115,9 @@ export interface IComponentInstance {
     targetId?: string; // list target ID (formId for type=form or modelId for type=model)
     fieldId?: string; // bind fieldId
     pageId?: string; // dataSource 中数据对象提取
-    features?: IFeatures;
+    features?: IFeatures; // extension features
+
+    search?: ISearch[];
   };
   events?: Record<EventAction, IEvent[]>;
   parent?: string;
@@ -149,7 +160,7 @@ export interface IComponentProps {
   dataSource?: IDataSource;
   isEdit: boolean;
   variables: Record<string, any>;
-  childrenRender?: (props: Record<string, any>, index: number) => ReactDOM;
+  childrenRender: (props: Record<string, any>, index: number) => ReactDOM;
   onSelectComponent: () => void;
   // for data source value change (eg: form data)
   onChangeValue: (value: any, keyPath?: string | null) => void;