+++ packages/mtbird-ui/src/index.ts
@@ -1,5 +1,8 @@
-export { default as List } from './components/List';
 export { default as SchemaSelect } from './components/SchemaSelect';
+export { default as SchemaInput } from './components/SchemaInput';
+
+export { default as List } from './components/List';
 export { default as CollapsePanel } from './components/CollapsePanel';
 export { default as FormItem } from './components/FormItem';
 export { default as Upload } from './components/Upload';
+export { default as EditableText } from './components/EditableText';